multi-agent/integrations/langchain_core/admin/models.py
"""
Model policy management.
"""

import logging
from typing import Dict, Any, List, Optional
from google.cloud import firestore

from ..config import get_config

logger = logging.getLogger(__name__)


class ModelPolicyManager:
    """
    Manages model policies for users and brands.
    """
    
    def __init__(self, db=None):
        """Initialize manager."""
        self.config = get_config()
        self.db = db
        
        if not self.db and self.config.firestore_project:
            self.db = firestore.Client(project=self.config.firestore_project)
    
    def get_policy(
        self,
        scope: str = "global",
        scope_id: Optional[str] = None
    ) -> Dict[str, Any]:
        """Get model policy."""
        # TODO: Implement policy retrieval from Firestore
        return {
            "provider": self.config.lc_provider,
            "model": self.config.lc_model,
            "temperature": self.config.lc_temperature,
            "max_tokens": self.config.lc_max_tokens
        }
    
    def update_policy(
        self,
        scope: str,
        scope_id: Optional[str],
        policy: Dict[str, Any]
    ) -> bool:
        """Update model policy."""
        # TODO: Implement policy update
        return True


# Global manager instance
_manager: Optional[ModelPolicyManager] = None


def get_policy_manager() -> ModelPolicyManager:
    """Get the global policy manager."""
    global _manager
    if _manager is None:
        _manager = ModelPolicyManager()
    return _manager