multi-agent/integrations/langchain_core/requirements.txt
# LangChain Core Requirements
# Install with: pip install -r requirements.txt -c ../../../constraints.txt

# Core LangChain
langchain>=0.3.14
langgraph>=0.2.55
langchain-core>=0.3.28
langchain-community>=0.3.14
langchain-text-splitters>=0.3.9

# LLM Providers
langchain-openai>=0.2.10
langchain-anthropic>=0.3.6
langchain-google-genai>=2.0.7

# Vector Stores
faiss-cpu>=1.8.0.post1

# Embeddings
sentence-transformers>=2.7.0

# Document Processing
unstructured>=0.14.8
unstructured[md]>=0.14.8

# Core Dependencies
tiktoken>=0.7.0
httpx>=0.28.1
tenacity>=8.3.0
pydantic>=2.11.7
pydantic-settings>=2.10.1

# Google Cloud
google-cloud-firestore>=2.21.0
google-cloud-secret-manager>=2.16.0

# CLI
typer>=0.15.1
rich>=13.9.4

# Scientific
numpy>=1.26.4
scipy>=1.13.1

# Testing
pytest>=8.2.2
pytest-asyncio>=0.23.7
pytest-mock>=3.14.0